Optimize `xla::HloSharding::PartialTile` by making the `get_group_id` lambda faster.
Currently the for-loop has a loop-carried dependency of 4 cycles (1x mul, 1x add) on the register for the `group_id` variable. We can precompute the strides for each index and directly multiply with it. Doing so reduces the loop-carried dependency to 1 cycle (1x add) on the register for the `group_id` variable.
- https://godbolt.org/z/oqvbzM9dP
- Analysis mode diff: [Before: 4 cycles / iteration](https://godbolt.org/z/ox9sfG8sW) vs [After: 1.3 cycles / iteration](https://godbolt.org/z/YEMTTvMY5)
- Bottleneck analysis for the "before" assembly: https://godbolt.org/z/jY6efhzY3

PiperOrigin-RevId: 748604884